% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsl_bear_river_refuge}
\alias{gsl_bear_river_refuge}
\title{Bear River example refuge boundary}
\format{
A packed \code{terra::SpatVector} polygon.
}
\source{
FWS Interest simplified authoritative boundary.
}
\usage{
gsl_bear_river_refuge
}
\description{
The cropped Bear River Migratory Bird Refuge boundary corresponding to
\code{gsl_bear_river_water}.
}
\details{
This object is a packed \code{terra::SpatVector}. Use \code{terra::vect()} to unpack it.
}
\examples{
if (requireNamespace("terra", quietly = TRUE)) {
  data(gsl_bear_river_refuge)
  refuge <- terra::vect(gsl_bear_river_refuge)
  refuge
}
}
\keyword{datasets}
