% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\docType{package}
\name{timescapes-package}
\alias{timescapes}
\alias{timescapes-package}
\title{Binary Timescape Metrics}
\description{
\code{timescapes} calculates mathematical summaries of binary conditions through
time, either for one vector or independently for every cell in a raster time
series.
}
\author{
\strong{Maintainer}: Logan Hysen \email{hysenlog@msu.edu} [copyright holder]

Authors:
\itemize{
  \item Logan Hysen \email{hysenlog@msu.edu} [copyright holder]
  \item Ho Yi Wan \email{wanhoyi@ufl.edu} [copyright holder]
}

}
\keyword{internal}
