% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_basic.R
\name{ts_aggregation}
\alias{ts_aggregation}
\title{Temporal aggregation}
\usage{
ts_aggregation(x)
}
\arguments{
\item{x}{Binary time series (0/1)}
}
\value{
Numeric (0–1)
}
\description{
Temporal aggregation
}
