% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_basic.R
\name{ts_autocorr}
\alias{ts_autocorr}
\title{Lag-1 temporal autocorrelation}
\usage{
ts_autocorr(x)
}
\arguments{
\item{x}{Numeric time series}
}
\value{
Numeric
}
\description{
Lag-1 temporal autocorrelation
}
