% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_data.R
\name{ts_benchmark}
\alias{ts_benchmark}
\title{Create benchmark temporal scenarios}
\usage{
ts_benchmark(n = 100)
}
\arguments{
\item{n}{Integer length of time series}
}
\value{
List of named binary vectors
}
\description{
Create benchmark temporal scenarios
}
