% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{ts_calculate}
\alias{ts_calculate}
\title{Calculate temporal metrics}
\usage{
ts_calculate(x, metrics = "all")
}
\arguments{
\item{x}{A binary time series (vector)}

\item{metrics}{Character vector of metrics (default = all)}
}
\value{
Named numeric vector
}
\description{
Calculate temporal metrics
}
