% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_structure.R
\name{ts_core_time_index}
\alias{ts_core_time_index}
\title{Core time index}
\usage{
ts_core_time_index(x, core_buffer = 1L)
}
\arguments{
\item{x}{Binary time series (0/1)}

\item{core_buffer}{Integer buffer}
}
\value{
Numeric (0–1)
}
\description{
Core time index
}
