% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_structure.R
\name{ts_duration_sd}
\alias{ts_duration_sd}
\title{Duration variability (SD of run lengths)}
\usage{
ts_duration_sd(x)
}
\arguments{
\item{x}{Binary time series (0/1)}
}
\value{
Numeric
}
\description{
Duration variability (SD of run lengths)
}
