% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_basic.R
\name{ts_edge_density}
\alias{ts_edge_density}
\title{Temporal edge density}
\usage{
ts_edge_density(x)
}
\arguments{
\item{x}{Binary time series (0/1)}
}
\value{
Numeric
}
\description{
Temporal edge density
}
