% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{ts_list_metrics}
\alias{ts_list_metrics}
\title{List available binary timescape metrics}
\usage{
ts_list_metrics()
}
\value{
A character vector of exact metric names accepted by \code{\link[=ts_metric]{ts_metric()}}
and \code{\link[=ts_raster_metric]{ts_raster_metric()}}.
}
\description{
List available binary timescape metrics
}
\examples{
ts_list_metrics()

}
