% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{ts_metric}
\alias{ts_metric}
\title{Calculate a single named timescape metric}
\usage{
ts_metric(x, metric)
}
\arguments{
\item{x}{Binary time series (0/1)}

\item{metric}{Metric name}
}
\value{
Numeric value
}
\description{
Calculate a single named timescape metric
}
