% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{ts_metrics}
\alias{ts_metrics}
\title{Calculate multiple timescape metrics}
\usage{
ts_metrics(x, metrics = names(.ts_registry))
}
\arguments{
\item{x}{Binary time series (0/1)}

\item{metrics}{Character vector of metric names}
}
\value{
Named numeric vector
}
\description{
Calculate multiple timescape metrics
}
