% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_structure.R
\name{ts_n_core_periods}
\alias{ts_n_core_periods}
\title{Number of core periods}
\usage{
ts_n_core_periods(x, core_buffer = 1L)
}
\arguments{
\item{x}{Binary time series (0/1)}

\item{core_buffer}{Integer buffer}
}
\value{
Integer
}
\description{
Number of core periods
}
