% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_structure.R
\name{ts_n_periods}
\alias{ts_n_periods}
\title{Number of occupancy periods}
\usage{
ts_n_periods(x)
}
\arguments{
\item{x}{Binary time series (0/1)}
}
\value{
Integer
}
\description{
Number of occupancy periods
}
